app/MVP/requirements.txt
# FastAPI框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis
redis==5.0.1
aioredis==2.0.1

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# AI/LLM
openai==1.3.7
anthropic==0.7.7
langchain==0.0.340
llama-index==0.9.15
transformers==4.35.2

# 工具库
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
aiofiles==23.2.1

# 日志和监控
python-json-logger==2.0.7
prometheus-client==0.19.0

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# 开发工具
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0